Fleshed out a few minor things.
This is getting to the point where we are going to need a system for defining types of relations, ontologies (collections of types
of relations), and so on in the general case.  There's sort of a slippery slope here, because basically ALL metadata defines a
relation, and we have to decide how we want to do it.

Categories are a related problem, as are, at a more abstract level, the questions of embedding sublanguages such as SVG or
whatever.